chore: separate routes from main.py to routes.py
Utilized flask global variable 'g' and Blueprint for middleware
